move web3 dir & remove $api on useBalance
